fix the problems
